Create sintax to estimate GPL parameter
because of the GPL linear model parameter will use the maximum likelihood estimation, where iteration will be carried out to obtain the parameter estimate, an initial iteration value is required. The recommended initial value for iteration is the estimated GPL parameter from the label data.
This syntax will be used to obtain the GPL distribution parameters from the label data.